refactor testing code, workflow
